+++ plugins/main/public/components/agents/syscollector/columns/netaddr-columns.ts
@@ -1,7 +1,0 @@
-export const netaddrColumns = [
-  { id: 'network.name' },
-  { id: 'agent.host.ip' },
-  { id: 'network.netmask' },
-  { id: 'network.protocol' },
-  { id: 'network.broadcast' },
-];