+++ plugins/main/public/components/agents/syscollector/columns/netiface-columns.ts
@@ -1,7 +1,0 @@
-export const netifaceColumns = [
-  { id: 'observer.ingress.interface.name' },
-  { id: 'host.mac' },
-  { id: 'observer.ingress.interface.state' },
-  { id: 'observer.ingress.interface.mtu' },
-  { id: 'observer.ingress.interface.type' },
-];