+++ plugins/main/public/components/agents/syscollector/columns/windows-updates-columns.ts
@@ -1,1 +1,0 @@
-export const windowsUpdatesColumns = [{ id: 'package.hotfix.name' }];