+++ plugins/main/public/components/common/data-source/pattern/system-inventory-network-settings/index.ts
@@ -1,2 +1,0 @@
-export * from './data-source-repository';
-export * from './data-source';