+++ plugins/main/public/components/overview/it-hygiene/dashboards/dashboard-panels.ts
@@ -1,0 +1,121 @@
+import { DashboardPanelState } from '../../../../../../../../src/plugins/dashboard/public/application';
+import { EmbeddableInput } from '../../../../../../../../src/plugins/embeddable/public';
+
+const getVisState = (indexPatternId: string) => {
+  return {
+    id: 'vis-packages',
+    title: 'vis-packages',
+    type: 'pie',
+    params: {
+      type: 'pie',
+      addTooltip: true,
+      addLegend: true,
+      legendPosition: 'right',
+      isDonut: true,
+      labels: {
+        show: false,
+        values: true,
+        last_level: true,
+        truncate: 100,
+      },
+    },
+    data: {
+      searchSource: {
+        query: {
+          language: 'kuery',
+          query: '',
+        },
+        filter: [],
+        index: indexPatternId,
+      },
+      references: [
+        {
+          name: 'kibanaSavedObjectMeta.searchSourceJSON.index',
+          type: 'index-pattern',
+          id: indexPatternId,
+        },
+      ],
+      aggs: [
+        {
+          id: '1',
+          enabled: true,
+          type: 'count',
+          params: {},
+          schema: 'metric',
+        },
+        {
+          id: '2',
+          enabled: true,
+          type: 'filters',
+          params: {
+            filters: [
+              {
+                input: {
+                  query: 'package.name: *',
+                  language: 'kuery',
+                },
+                label: '',
+              },
+            ],
+          },
+          schema: 'segment',
+        },
+        {
+          id: '3',
+          enabled: true,
+          type: 'terms',
+          params: {
+            field: 'agent.id',
+            orderBy: '1',
+            order: 'desc',
+            size: 5,
+            otherBucket: false,
+            otherBucketLabel: 'Other',
+            missingBucket: false,
+            missingBucketLabel: 'Missing',
+          },
+          schema: 'segment',
+        },
+      ],
+    },
+  };
+};
+
+export const getDashboardPanels = (
+  indexPatternId: string,
+): {
+  [panelId: string]: DashboardPanelState<
+    EmbeddableInput & { [k: string]: unknown }
+  >;
+} => {
+  return {
+    '1': {
+      gridData: {
+        w: 16,
+        h: 12,
+        x: 0,
+        y: 0,
+        i: '1',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '1',
+        savedVis: getVisState(indexPatternId),
+      },
+    },
+    '7': {
+      gridData: {
+        w: 16,
+        h: 12,
+        x: 16,
+        y: 0,
+        i: '7',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '7',
+        savedVis: getVisState(indexPatternId),
+      },
+    },
+  };
+};