+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Support for Wazuh 4.12.0
 - Add setting to limit the number of rows in csv reports [#7182](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7182)
 - Add new `vulnerability.scanner.reference` field which contains the CTI reference of the vulnerability [#7306](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7306)
+- Added `It Hygiene` application [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
+- Added hardware and system information to agent overview [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
 
 ### Changed
 
@@ -28,6 +30,11 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed logging UI logs due to logger undefined property [#7267](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7267)
 - Fixed TOP-5-SO filter management in Endpoints > Summary [#7278](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7278)
 - Fixed CSV export not filtering by timerange [#7304](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7304)
+
+### Removed
+
+- Removed `Inventory data` from agent overview [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
+- Removed `GET /api/syscollector` API endpoint [#7368](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7368)
 
 ## Wazuh v4.11.1 - OpenSearch Dashboards 2.16.0 - Revision 01